Convert Home Page to not use useSelector and useDispatch hook, use mapStateToProps & mapDispatchToProps with connect...... Separate all actions of  into separate files and export it from index.j (main file of actions)s
